+++ trane/core/prediction_problem_generator.py
@@ -5,10 +5,10 @@ from tqdm.notebook import tqdm
 
 from trane.core.prediction_problem import PredictionProblem
 from trane.core.utils import _parse_table_meta, get_semantic_tags
-from trane.ops import aggregation_ops as agg_ops
-from trane.ops import filter_ops
+from trane.ops.aggregation_ops import CountAggregationOp
 from trane.ops.filter_ops import AllFilterOp
 from trane.ops.threshold_functions import get_k_most_frequent
+from trane.ops.utils import get_aggregation_ops, get_filter_ops
 from trane.typing.column_schema import ColumnSchema
 from trane.typing.inference import infer_table_meta
 from trane.typing.logical_types import (
@@ -116,15 +116,15 @@ class PredictionProblemGenerator:
 
         possible_ops = []
         for agg, filter_ in itertools.product(
-            agg_ops.AGGREGATION_OPS,
-            filter_ops.FILTER_OPS,
+            get_aggregation_ops(),
+            get_filter_ops(),
         ):
             filter_columns = all_columns
-            if filter_ == "AllFilterOp":
+            if filter_ == AllFilterOp:
                 filter_columns = [None]
 
             agg_columns = all_columns
-            if agg == "CountAggregationOp":
+            if agg == CountAggregationOp:
                 agg_columns = [None]
             for filter_col, agg_col in itertools.product(
                 filter_columns,
@@ -143,10 +143,10 @@ class PredictionProblemGenerator:
             position=pbar_position,
         ):
             all_attempts += 1
-            ag_col, filter_col, agg_op_name, filter_op_name = op_col_combo
+            ag_col, filter_col, agg_op_obj, filter_op_obj = op_col_combo
 
-            agg_op_obj = getattr(agg_ops, agg_op_name)(ag_col)  # noqa
-            filter_op_obj = getattr(filter_ops, filter_op_name)(filter_col)  # noqa
+            agg_op_obj = agg_op_obj(ag_col)
+            filter_op_obj = filter_op_obj(filter_col)
 
             # Note: the order of the operations matters, the filter operation must be first
             operations = [filter_op_obj, agg_op_obj]