+++ trane/ops/op_base.py
@@ -33,6 +33,30 @@ class OpBase(object):
     def __call__(self, dataslice):
         return self.label_function(dataslice)
 
+    def __lt__(self, other):
+        return (type(self).__name__) < (type(other).__name__)
+
+    def __le__(self, other):
+        return (type(self).__name__) <= (type(other).__name__)
+
+    def __gt__(self, other):
+        return (type(self).__name__) > (type(other).__name__)
+
+    def __ge__(self, other):
+        return (type(self).__name__) >= (type(other).__name__)
+
+    def __hash__(self):
+        return hash((type(self).__name__, self.column_name))
+
+    def __repr__(self):
+        return type(self).__name__
+
+    def __eq__(self, other):
+        """Overrides the default implementation"""
+        if isinstance(self, other.__class__):
+            return self.__dict__ == other.__dict__
+        return False
+
     def label_function(self, dataslice):
         raise NotImplementedError
 
@@ -76,20 +100,3 @@ class OpBase(object):
                 best_threshold = unique_val
         self.set_parameters(threshold=original_threshold)
         return best_threshold
-
-    def __hash__(self):
-        return hash((type(self).__name__, self.column_name))
-
-    # def __repr__(self):
-    #     hyper_param_str = ",".join(
-    #         [str(x) for x in list(self.hyper_parameter_settings.values())],
-    #     )
-    #     if len(hyper_param_str) > 0:
-    #         hyper_param_str = "@" + hyper_param_str
-    #     return "%s(%s%s)" % (type(self).__name__, self.column_name, hyper_param_str)
-
-    def __eq__(self, other):
-        """Overrides the default implementation"""
-        if isinstance(self, other.__class__):
-            return self.__dict__ == other.__dict__
-        return False