+++ docs/changelog.md
@@ -6,10 +6,12 @@ Changelog
 v0.5.0 ()
 =====================
 * Enhancements
-    * Add ExistsAggregationOp  [#96][#96]
+    * Add ``ExistsAggregationOp`` [#96][#96]
+    * Add `get_aggregation_ops` and `get_filter_ops` functions [#98][#98]
 * Fixes
 
     [#96]: <https://github.com/trane-dev/Trane/pull/96>
+    [#98]: <https://github.com/trane-dev/Trane/pull/98>
 
 v0.4.0 (July 8, 2023)
 =====================